Fix a build break caused by a type mismatch on 64-bit build targets.
Review URL: https://chromiumcodereview.appspot.com//10871096

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@11430 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/pages.cc
@@ -532,7 +532,7 @@ void PageSpaceController::EvaluateGarbageCollection(
                                                      desired_utilization_);
       intptr_t growth_in_bytes = Utils::RoundUp(growth_target - in_use_after,
                                                 PageSpace::kPageSize);
-      intptr_t growth_in_pages = growth_in_bytes / PageSpace::kPageSize;
+      int growth_in_pages = growth_in_bytes / PageSpace::kPageSize;
       grow_heap_ = Utils::Maximum(growth_in_pages, heap_growth_rate_);
     } else {
       grow_heap_ = heap_growth_rate_;
